dev sale

+++ database/migrations/1523502464205_goods_group_schema.js
@@ -8,7 +8,7 @@ class GoodsGroupSchema extends Schema {
       table.increments('ni_id')
       table.integer('goods_id', 12)
       table.string('group_depict', 255)
-      table.integer('group_price', 32)
+      table.float('group_price', 32)
       table.integer('group_instock', 32)
       table.string('group_thumb', 64)
       table.integer('group_status', 1)
